requirements.txt
apiclient==1.0.4
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
google-api-core==2.0.0
google-api-python-client==2.17.0
google-auth==2.0.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
